requirements.txt
crewai
langchain-google-genai
python-dotenv
crewai-tools
langchain-openai
